+++ sample/src/main/java/me/panpf/adapter/sample/ui/RecyclerPagedSampleFragment.kt
@@ -9,6 +9,7 @@ import androidx.lifecycle.Observer
 import androidx.recyclerview.widget.LinearLayoutManager
 import kotlinx.android.synthetic.main.fm_recycler.*
 import me.panpf.adapter.paged.AssemblyPagedListAdapter
+import me.panpf.adapter.paged.DiffableDiffCallback
 import me.panpf.adapter.sample.R
 import me.panpf.adapter.sample.item.GameItem
 import me.panpf.adapter.sample.item.HeaderItem
@@ -21,7 +22,7 @@ import me.panpf.arch.ktx.bindViewModel
 class RecyclerPagedSampleFragment : BaseFragment() {
 
     private val viewModel by bindViewModel(ListViewModel::class)
-    private val adapter = AssemblyPagedListAdapter<Any>().apply {
+    private val adapter = AssemblyPagedListAdapter<Any>(DiffableDiffCallback()).apply {
         addHeaderItem(HeaderItem.Factory(), "我是小额头呀！")
         addItemFactory(UserItem.Factory())
         addItemFactory(GameItem.Factory())